requirements.txt
ultralytics
sahi
opencv-python
requests
earthengine-api
staticmap
pillow
pandas
tensorrt